requirements.txt
kerykeion==5.5.3
openai>=1.0.0
pytz>=2024.2
geopy==2.4.1
timezonefinder==6.5.2
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
requests>=2.31.0
psycopg2-binary>=2.9.9
